# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recruitment Dashboard API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/resume/upload": {
      "post": {
        "tags": [
          "Resume"
        ],
        "summary": "Upload Resume",
        "operationId": "upload_resume_api_resume_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_resume_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadEnqueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/upload/bulk": {
      "post": {
        "tags": [
          "Resume"
        ],
        "summary": "Upload Resume Bulk",
        "operationId": "upload_resume_bulk_api_resume_upload_bulk_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_bulk_api_resume_upload_bulk_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUploadEnqueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/upload/url": {
      "post": {
        "tags": [
          "Resume"
        ],
        "summary": "Upload Resume By Url",
        "operationId": "upload_resume_by_url_api_resume_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeUrlUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadEnqueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/upload/url/bulk": {
      "post": {
        "tags": [
          "Resume"
        ],
        "summary": "Upload Resume Urls Bulk",
        "operationId": "upload_resume_urls_bulk_api_resume_upload_url_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeBulkUrlUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUploadEnqueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_resume_api_resume_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "requirement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_resume_upload_post"
      },
      "Body_upload_resume_bulk_api_resume_upload_bulk_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "requirement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Id"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_resume_bulk_api_resume_upload_bulk_post"
      },
...
      "BulkUploadEnqueueResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "status",
          "accepted",
          "rejected"
        ],
        "title": "BulkUploadEnqueueResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeBulkUrlUploadRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "minItems": 1,
            "title": "Urls"
          },
          "requirement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Id"
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "ResumeBulkUrlUploadRequest"
      },
      "ResumeUrlUploadRequest": {
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "requirement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Id"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ResumeUrlUploadRequest"
      },
...
      "UploadEnqueueResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "status",
          "accepted"
        ],
        "title": "UploadEnqueueResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}